requirements.txt
fastapi
uvicorn
langchain
langchain_community
langchain_core
langchain-groq
faiss-cpu
transformers
sentence-transformers
httpx
pillow
pytesseract
python-dotenv
